Move button components
